# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List schedule history for a device
      parameters:
        - name: deviceId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of schedule summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleSummary"
  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      summary: Get a schedule with its blocks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: deviceId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Schedule detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Manually edit a schedule's blocks
      description: Replaces the schedule's blocks with the provided list (manual editing).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScheduleInput"
      responses:
        "200":
          description: Updated schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Delete a schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: deviceId
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /schedules/{id}/revise:
    post:
      operationId: reviseSchedule
      tags: [schedules]
      summary: Ask AI to revise a schedule with a natural-language instruction
      description: >
        Takes an existing schedule and a free-text instruction (e.g. "move my
        study block later" or "add more breaks") and returns the revised
        schedule with updated blocks.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviseScheduleInput"
      responses:
        "200":
          description: Revised schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
...
  /schedules/generate:
    post:
      operationId: generateSchedule
      tags: [schedules]
      summary: Generate an AI-recommended schedule, or ask clarifying questions first
      description: >
        Takes a student's commitments and generates a recommended schedule.
        If the AI needs more information (e.g. dinner time, homework load) it
        returns status "needs_clarification" with a list of questions instead
        of a finished schedule. Re-call with the same draftId and answers to
        continue.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateScheduleInput"
      responses:
        "200":
          description: Either a completed schedule or a request for clarification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateScheduleResult"
...
components:
  schemas:
...
    DayOfWeek:
      type: string
      enum: [mon, tue, wed, thu, fri, sat, sun]
...
    UserPreferences:
      type: object
      properties:
        wakeTime:
          type: string
          nullable: true
          description: 'When the student usually wakes up (e.g. "7:00 AM").'
        bedTime:
          type: string
          nullable: true
          description: 'When the student usually goes to bed (e.g. "10:30 PM").'
        mealTimes:
          type: string
          nullable: true
          description: 'Usual meal times, free text (e.g. "breakfast 7:30, lunch 12, dinner 6:30").'
        notes:
          type: string
          nullable: true
          description: Anything else the AI should keep in mind when scheduling.
      required: []
...
    Task:
      type: object
      description: An assignment or task the student needs to work on.
      properties:
        title:
          type: string
        dueDate:
          type: string
          description: 'When it is due, free text (e.g. "tomorrow", "Friday", "2026-07-10").'
        estimatedMinutes:
          type: integer
          nullable: true
          description: Rough estimate of how long the task will take, in minutes.
        notes:
          type: string
          nullable: true
      required:
        - title
        - dueDate
    ScheduleBlockCategory:
      type: string
      enum: [class, extracurricular, routine, homework, break, free]
    ScheduleBlock:
      type: object
      properties:
        id:
          type: string
        day:
          $ref: "#/components/schemas/DayOfWeek"
        startTime:
          type: string
        endTime:
          type: string
        title:
          type: string
        category:
          $ref: "#/components/schemas/ScheduleBlockCategory"
        notes:
          type: string
          nullable: true
      required:
        - id
        - day
        - startTime
        - endTime
        - title
        - category
    ScheduleStatus:
      type: string
      enum: [needs_clarification, complete]
    ScheduleScope:
      type: string
      enum: [day, week]
    ScheduleSummary:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        scope:
          $ref: "#/components/schemas/ScheduleScope"
        status:
          $ref: "#/components/schemas/ScheduleStatus"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - scope
        - status
        - createdAt
    Schedule:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        scope:
          $ref: "#/components/schemas/ScheduleScope"
        status:
          $ref: "#/components/schemas/ScheduleStatus"
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleBlock"
        clarifyingQuestions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - deviceId
        - scope
        - status
        - blocks
        - clarifyingQuestions
        - createdAt
    ScheduleBlockInput:
      type: object
      properties:
        id:
          type: string
          description: "Stable block id. Include for existing blocks so Google Calendar sync mapping is preserved; omit for newly added blocks and the server will assign one."
        day:
          $ref: "#/components/schemas/DayOfWeek"
        startTime:
          type: string
        endTime:
          type: string
        title:
          type: string
        category:
          $ref: "#/components/schemas/ScheduleBlockCategory"
        notes:
          type: string
          nullable: true
      required:
        - day
        - startTime
        - endTime
        - title
        - category
    UpdateScheduleInput:
      type: object
      properties:
        deviceId:
          type: string
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/ScheduleBlockInput"
      required:
        - deviceId
        - blocks
    ReviseScheduleInput:
      type: object
      properties:
        deviceId:
          type: string
        instruction:
          type: string
          description: Natural-language description of the change the student wants.
      required:
        - deviceId
        - instruction
    ClarificationAnswer:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
      required:
        - question
        - answer
    GenerateScheduleInput:
      type: object
      properties:
        deviceId:
          type: string
        scope:
          $ref: "#/components/schemas/ScheduleScope"
        draftId:
          type: string
          nullable: true
          description: Pass the id returned from a prior needs_clarification response to continue that draft
        answers:
          type: array
          items:
            $ref: "#/components/schemas/ClarificationAnswer"
        tasks:
          type: array
          description: Assignments/tasks the student needs time for in this schedule.
          items:
            $ref: "#/components/schemas/Task"
        preferences:
          $ref: "#/components/schemas/UserPreferences"
          description: >
            The student's scheduling preferences (sleep, meals, extra notes).
            When provided, the AI should use these instead of asking
            clarifying questions.
      required:
        - deviceId
        - scope
...
    GenerateScheduleResult:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/ScheduleStatus"
        questions:
          type: array
          items:
            type: string
        schedule:
          $ref: "#/components/schemas/Schedule"
      required:
        - id
        - status
...