# lib/api-zod/src/generated/types/generateScheduleInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClarificationAnswer } from './clarificationAnswer';
import type { ScheduleScope } from './scheduleScope';
import type { Task } from './task';
import type { UserPreferences } from './userPreferences';

export interface GenerateScheduleInput {
  deviceId: string;
  scope: ScheduleScope;
  /** Pass the id returned from a prior needs_clarification response to continue that draft */
  draftId?: string | null;
  answers?: ClarificationAnswer[];
  /** Assignments/tasks the student needs time for in this schedule. */
  tasks?: Task[];
  /** The student's scheduling preferences (sleep, meals, extra notes). When provided, the AI should use these instead of asking clarifying questions. */
  preferences?: UserPreferences;
}